feat(filters): rename
